requirements.txt
APScheduler==3.5.3
Flask==1.0.2
influxdb==4.1.1
paho-mqtt==1.3.0
pika==0.10.0
protobuf==3.4.0
pymongo==3.7.2
requests==2.21.0
requests-oauthlib==1.2.0
virtualenv==15.0.1
Werkzeug==0.14.1
celery[redis]
xmltodict==0.12.0
